refactor: Layout 분리로 상태 변경 시 레이아웃이 리랜더링 되지 않도록 변경
