+++ src/pages/Lotto.tsx
@@ -1,6 +1,3 @@
-import * as Styled from './Lotto.style';
-import { Footer } from '../components/Footer/Footer';
-import { Header } from '../components/Header/Header';
 import { InputOrderAmount } from '../components/InputOrderAmount/InputOrderAmount';
 import { useBuyLotto } from '../hooks/useBuyLotto';
 import { ShowMyTickets } from '../components/ShowMyTickets/ShowMyTickets';
@@ -9,6 +6,7 @@ import { useReducer } from 'react';
 import { StatisticsModal } from '../components/StatisticsModal/StatisticsModal';
 import { useInputWinNumbers } from '../hooks/useInputWinNumbers';
 import { useGetLottoStatistics } from '../hooks/useGetLottoStatistics';
+import { Layout } from '../components/Layout/Layout';
 
 export const Lotto = () => {
   const { amountInput, setAmountInput, buyLotto, lottoTickets } = useBuyLotto();
@@ -26,24 +24,19 @@ export const Lotto = () => {
   };
 
   return (
-    <Styled.Page>
-      <Header />
-      <Styled.Main>
-        <Styled.Title>🎱 내 번호 당첨 확인 🎱</Styled.Title>
-        <InputOrderAmount amountInput={amountInput} setAmountInput={setAmountInput} buyLotto={buyLotto} />
-        {lottoTickets.length > 0 && (
-          <>
-            <ShowMyTickets lottoTickets={lottoTickets} />
-            <InputWinNumbers
-              {...rest}
-              compareTicketAndWinNumbers={compareTicketAndWinNumbers}
-              handleModalState={handleModalState}
-            />
-          </>
-        )}
-      </Styled.Main>
-      <Footer />
+    <Layout>
+      <InputOrderAmount amountInput={amountInput} setAmountInput={setAmountInput} buyLotto={buyLotto} />
+      {lottoTickets.length > 0 && (
+        <>
+          <ShowMyTickets lottoTickets={lottoTickets} />
+          <InputWinNumbers
+            {...rest}
+            compareTicketAndWinNumbers={compareTicketAndWinNumbers}
+            handleModalState={handleModalState}
+          />
+        </>
+      )}
       {isModalOpen && <StatisticsModal lottoStatistics={lottoStatistics} onClose={handleModalState} />}
-    </Styled.Page>
+    </Layout>
   );
 };
